+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Avoid direct jakarta.inject annotations usages to work properly when jakarta.inject 1
+    is forced in classpath (by dropwizard BOM)
+
 ### 6.2.0 (2023-11-28)
 * Update to dropwizard 3.0.4
 * Add jsr305 as direct dependency to simplify usage `@javax.annotation.Nullable`